refactor to separate accept mcms ownership sequence
